+++ vagrant_workspace/Vagrantfile
@@ -1,5 +1,5 @@
 BOX = "ubuntu/xenial64"
-NUMBER_OF_NODES = 2
+NUMBER_OF_NODES = 1
 
 Vagrant.configure("2") do |config|
   config.vm.define "node-1" do |nodeConfig|
